requirements.txt
langchain
langchain-openai
openai
pinecone-client
python-dotenv
pypdf
flask
flask-cors
-e .